added doxygens

+++ src/crepe/manager/LoopTimerManager.cpp
@@ -81,7 +81,7 @@ double LoopTimerManager::get_lag() const {
 double LoopTimerManager::get_fixed_delta_time() const {
 	return this->fixed_delta_time.count() * this->time_scale;
 }
-void LoopTimerManager::set_fixed_delta_time(int seconds) {
+void LoopTimerManager::set_fixed_delta_time(double seconds) {
 	this->fixed_delta_time = std::chrono::duration<double>(seconds);
 }
